Move Timeline to different directory
